+++ src/go.mod
@@ -3,3 +3,5 @@ module AkwardSilents
 go 1.20
 
 require golang.org/x/net v0.19.0
+
+require github.com/joho/godotenv v1.5.1 // indirect
