+++ src/pkg/service/server/Readloop.go
@@ -1,7 +1,7 @@
 package server
 
 import (
-	"AkwardSilents/pkg/service/handler"
+	"AkwardSilents/pkg/service"
 	"fmt"
 	"golang.org/x/net/websocket"
 	"io"
@@ -19,10 +19,7 @@ func (s *Server) readLoopAccount(ws *websocket.Conn) {
 			continue
 		}
 		msg := buf[:n]
-		//TODO: Process the received message if needed
-		//fmt.Println("Received message:", string(msg))
-
-		_, err = ws.Write([]byte(handler.MessageHandlerAccount(msg)))
+		_, err = ws.Write([]byte(service.MessageHandlerAccount(msg))) //Process the received message
 		if err != nil {
 			return
 		}
@@ -41,7 +38,6 @@ func (s *Server) readLoopChat(ws *websocket.Conn) {
 			continue
 		}
 		msg := buf[:n]
-		//TODO: Process the received message if needed
 		fmt.Println("Received message:", string(msg))
 	}
 }
