+++ src/pkg/service/server/WebsocketHandler.go
@@ -5,15 +5,15 @@ import (
 	"golang.org/x/net/websocket"
 )
 
-func (s Server) HandleWSAccount(conn *websocket.Conn) {
+func (s *Server) HandleWSAccount(conn *websocket.Conn) {
 	fmt.Println("new incoming connection from client:", conn.RemoteAddr())
 	s.Conns[conn] = true
 
 	s.readLoopAccount(conn)
 }
 
-func (s Server) HandleWSChat(conn *websocket.Conn) {
-	fmt.Println("new incoming connection from client:", conn.RemoteAddr())
+func (s *Server) HandleWSChat(conn *websocket.Conn) {
+	fmt.Println("New Connect:", conn.RemoteAddr())
 	s.Conns[conn] = true
 
 	s.readLoopChat(conn)
